Fix part of #24717: Merge certificate assessment offering model into certificate-assessment.model
