+++ core/templates/components/certificate-assessment-offering-helper/certificate-offering-details.component.ts
@@ -30,7 +30,7 @@ import {
   ClassroomBackendApiService,
   ClassroomSummaryDict,
 } from 'domain/classroom/classroom-backend-api.service';
-import {CertificateAssessmentOfferingData} from 'domain/certificate-assessment/certificate-assessment-offering.model';
+import {CertificateAssessmentOfferingData} from 'domain/certificate-assessment/certificate-assessment.model';
 import './certificate-offering-details.component.css';
 
 interface CertificateOfferingDetailsFormData {
