+++ core/templates/components/certificate-assessment-offering-helper/certificate-offering-review-and-availability.component.ts
@@ -22,7 +22,7 @@ import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
 
 import {ClassroomBackendApiService} from 'domain/classroom/classroom-backend-api.service';
 import {CertificateAssessmentOfferingBackendApiService} from 'domain/certificate-assessment/certificate-assessment-offering-backend-api.service';
-import {CertificateAssessmentOfferingData} from 'domain/certificate-assessment/certificate-assessment-offering.model';
+import {CertificateAssessmentOfferingData} from 'domain/certificate-assessment/certificate-assessment.model';
 
 import './certificate-offering-review-and-availability.component.css';
 // Shape of one difficulty bucket returned by the validation API.
