+++ core/templates/domain/certificate-assessment/certificate-assessment.model.ts
@@ -13,7 +13,8 @@
 // limitations under the License.
 
 /**
- * @fileoverview Interfaces for frontend certificate assessment domain objects.
+ * @fileoverview Frontend models and interfaces for certificate assessment
+ * domain objects.
  */
 
 // The shape of each attempt summary mirrors the response returned by the
@@ -74,4 +75,199 @@ export interface RecommendedTopicStub {
   lessonCount: number;
   // Placeholder swatch shown instead of a topic thumbnail image.
   colorClass: string;
+}
+
+export interface CertificateAssessmentOfferingTopicData {
+  [topicId: string]: number;
+}
+
+export interface CertificateAssessmentOfferingBackendDict {
+  certificate_id: string;
+  title: string;
+  description: string;
+  classroom_id: string;
+  topic_ids: string[];
+  topic_data: CertificateAssessmentOfferingTopicData;
+  demonstrates: string[];
+  total_questions: number;
+  time_limit_in_minutes: number;
+  async_status: string;
+  version: number;
+}
+
+export interface AvailableCertificateAssessmentOfferingBackendDict {
+  certificate_id: string;
+  title: string;
+  attempt_status: string;
+}
+
+export class AvailableCertificateAssessmentOfferingData {
+  _certificateId: string;
+  _title: string;
+  _attemptStatus: string;
+
+  constructor(certificateId: string, title: string, attemptStatus: string) {
+    this._certificateId = certificateId;
+    this._title = title;
+    this._attemptStatus = attemptStatus;
+  }
+
+  get certificateId(): string {
+    return this._certificateId;
+  }
+
+  get title(): string {
+    return this._title;
+  }
+
+  get attemptStatus(): string {
+    return this._attemptStatus;
+  }
+
+  static createFromBackendDict(
+    availableCertificateAssessmentOfferingBackendDict: AvailableCertificateAssessmentOfferingBackendDict
+  ): AvailableCertificateAssessmentOfferingData {
+    return new AvailableCertificateAssessmentOfferingData(
+      availableCertificateAssessmentOfferingBackendDict.certificate_id,
+      availableCertificateAssessmentOfferingBackendDict.title,
+      availableCertificateAssessmentOfferingBackendDict.attempt_status
+    );
+  }
+}
+
+export class CertificateAssessmentOfferingData {
+  _certificateId: string;
+  _title: string;
+  _description: string;
+  _classroomId: string;
+  _topicData: CertificateAssessmentOfferingTopicData;
+  _totalQuestions: number;
+  _timeLimitInMinutes: number;
+  _demonstrates: string[];
+  _asyncStatus: string;
+  _version: number;
+
+  constructor(
+    certificateId: string,
+    title: string,
+    description: string,
+    classroomId: string,
+    topicData: CertificateAssessmentOfferingTopicData,
+    totalQuestions: number,
+    timeLimitInMinutes: number,
+    demonstrates: string[],
+    asyncStatus: string,
+    version: number
+  ) {
+    this._certificateId = certificateId;
+    this._title = title;
+    this._description = description;
+    this._classroomId = classroomId;
+    this._topicData = topicData;
+    this._totalQuestions = totalQuestions;
+    this._timeLimitInMinutes = timeLimitInMinutes;
+    this._demonstrates = demonstrates;
+    this._asyncStatus = asyncStatus;
+    this._version = version;
+  }
+
+  get certificateId(): string {
+    return this._certificateId;
+  }
+
+  get title(): string {
+    return this._title;
+  }
+
+  set title(title: string) {
+    this._title = title;
+  }
+
+  get description(): string {
+    return this._description;
+  }
+
+  set description(description: string) {
+    this._description = description;
+  }
+
+  get classroomId(): string {
+    return this._classroomId;
+  }
+
+  set classroomId(classroomId: string) {
+    this._classroomId = classroomId;
+  }
+
+  get topicData(): CertificateAssessmentOfferingTopicData {
+    return this._topicData;
+  }
+
+  set topicData(topicData: CertificateAssessmentOfferingTopicData) {
+    this._topicData = topicData;
+  }
+
+  get totalQuestions(): number {
+    return this._totalQuestions;
+  }
+
+  set totalQuestions(totalQuestions: number) {
+    this._totalQuestions = totalQuestions;
+  }
+
+  get timeLimitInMinutes(): number {
+    return this._timeLimitInMinutes;
+  }
+
+  set timeLimitInMinutes(timeLimitInMinutes: number) {
+    this._timeLimitInMinutes = timeLimitInMinutes;
+  }
+
+  get demonstrates(): string[] {
+    return this._demonstrates;
+  }
+
+  set demonstrates(demonstrates: string[]) {
+    this._demonstrates = demonstrates;
+  }
+
+  get asyncStatus(): string {
+    return this._asyncStatus;
+  }
+
+  get version(): number {
+    return this._version;
+  }
+
+  static createEmpty(): CertificateAssessmentOfferingData {
+    return new CertificateAssessmentOfferingData(
+      '',
+      '',
+      '',
+      '',
+      {},
+      0,
+      0,
+      [],
+      'Not_Ready',
+      0
+    );
+  }
+
+  static createFromBackendDict(
+    certificateAssessmentOfferingBackendDict: CertificateAssessmentOfferingBackendDict
+  ): CertificateAssessmentOfferingData {
+    return new CertificateAssessmentOfferingData(
+      certificateAssessmentOfferingBackendDict.certificate_id,
+      certificateAssessmentOfferingBackendDict.title,
+      certificateAssessmentOfferingBackendDict.description,
+      certificateAssessmentOfferingBackendDict.classroom_id,
+      certificateAssessmentOfferingBackendDict.topic_data,
+      certificateAssessmentOfferingBackendDict.total_questions,
+      certificateAssessmentOfferingBackendDict.time_limit_in_minutes,
+      certificateAssessmentOfferingBackendDict.demonstrates,
+      certificateAssessmentOfferingBackendDict.async_status,
+      certificateAssessmentOfferingBackendDict.version
+    );
+  }
 }
